refactor: define dedupeExtensions() locally in resolveVikeConfigInternal
dedupeExtensions() is only used by resolveConfigValueSources(), so define it as
a local helper there (reading the extension `name` via the local getConfVal())
instead of exporting it from getPlusFilesByLocationId. getPlusFilesByLocationId
and assertExtensions no longer carry any extension-dedup code.
